Add ParsedError. Check request headers, CSeq num.
1. ParsedError is an Error suggesting a status code and reason phrase with
which the application should respond, if the message was a Request. A
ParsedError also contains a parsed message in which malformed headers are left
unparsed.

2. When parsing a Request, ensure that To, From, CSeq, Call-ID, Max-Forwards,
and Via headers are present and well-formed. Throw a ParsedError if they
aren't.  See http://tools.ietf.org/html/rfc3261#section-8.1.1

3. When parsing any SIP message, ensure that the CSeq sequence number is less
than 2**32. Throw a ParsedError if it isn't. See
http://tools.ietf.org/html/rfc3261#section-20.16.